move parsers back to ssd
